+++ src/tests/datastore-tests.ts
@@ -23,11 +23,11 @@ import {
 } from '../datastore/common';
 import { parseDbEvent } from '../api/controllers/db-controller';
 import * as assert from 'assert';
-import { I32_MAX } from '../helpers';
-import { intCV, serializeCV } from '@stacks/transactions';
 import { PgWriteStore } from '../datastore/pg-write-store';
 import { cycleMigrations, runMigrations } from '../datastore/migrations';
 import { getPostgres, PgSqlClient } from '../datastore/connection';
+import { bnsNameCV, I32_MAX } from '../helpers';
+import { ChainID, intCV, serializeCV } from '@stacks/transactions';
 
 function testEnvVars(
   envVars: Record<string, string | undefined>,
@@ -3587,25 +3587,6 @@ describe('postgres datastore', () => {
       execution_cost_write_count: 0,
       execution_cost_write_length: 0,
     };
-    const block3: DbBlock = {
-      block_hash: '0x33',
-      index_block_hash: '0xcc',
-      parent_index_block_hash: block2.index_block_hash,
-      parent_block_hash: block2.block_hash,
-      parent_microblock_hash: '0x00',
-      block_height: 3,
-      burn_block_time: 1234,
-      burn_block_hash: '0x1234',
-      burn_block_height: 123,
-      miner_txid: '0x4321',
-      canonical: true,
-      parent_microblock_sequence: 0,
-      execution_cost_read_count: 0,
-      execution_cost_read_length: 0,
-      execution_cost_runtime: 0,
-      execution_cost_write_count: 0,
-      execution_cost_write_length: 0,
-    };
 
     const minerReward1: DbMinerReward = {
       ...block1,
@@ -3745,12 +3726,25 @@ describe('postgres datastore', () => {
           stxLockEvents: [stxLockEvent2],
           stxEvents: [],
           ftEvents: [],
-          nftEvents: [],
+          nftEvents: [
+            {
+              event_type: DbEventTypeId.NonFungibleTokenAsset,
+              asset_event_type_id: DbAssetEventTypeId.Mint,
+              value: bnsNameCV('xyz.abc'),
+              asset_identifier: 'SP000000000000000000002Q6VF78.bns::names',
+              recipient: 'ST5RRX0K27GW0SP3GJCEMHD95TQGJMKB7G9Y0X1ZA',
+              tx_id: tx2.tx_id,
+              tx_index: tx2.tx_index,
+              block_height: 2,
+              event_index: 1,
+              canonical: true,
+            },
+          ],
           contractLogEvents: [],
           smartContracts: [],
           names: [
             {
-              name: 'xyz',
+              name: 'xyz.abc',
               address: 'ST5RRX0K27GW0SP3GJCEMHD95TQGJMKB7G9Y0X1ZA',
               namespace_id: 'abc',
               registered_at: 2,
@@ -3814,7 +3808,11 @@ describe('postgres datastore', () => {
       ]
     );
 
-    let name = await db.getName({ name: 'xyz', includeUnanchored: false });
+    let name = await db.getName({
+      name: 'xyz.abc',
+      includeUnanchored: false,
+      chainId: ChainID.Mainnet,
+    });
     assert(name.found);
     expect(name.result.canonical).toBe(true);
     expect(name.result.index_block_hash).toBe(block2.index_block_hash);
@@ -3829,6 +3827,25 @@ describe('postgres datastore', () => {
     expect(subdomain.result.canonical).toBe(true);
     expect(subdomain.result.index_block_hash).toBe(block2.index_block_hash);
 
+    const block3: DbBlock = {
+      block_hash: '0x33',
+      index_block_hash: '0xcc',
+      parent_index_block_hash: block2.index_block_hash,
+      parent_block_hash: block2.block_hash,
+      parent_microblock_hash: '',
+      block_height: 3,
+      burn_block_time: 1234,
+      burn_block_hash: '0x1234',
+      burn_block_height: 123,
+      miner_txid: '0x4321',
+      canonical: true,
+      parent_microblock_sequence: 0,
+      execution_cost_read_count: 0,
+      execution_cost_read_length: 0,
+      execution_cost_runtime: 0,
+      execution_cost_write_count: 0,
+      execution_cost_write_length: 0,
+    };
     await db.update({ block: block3, microblocks: [], minerRewards: [], txs: [] });
 
     const block2b: DbBlock = {
@@ -3903,22 +3920,35 @@ describe('postgres datastore', () => {
           stxLockEvents: [],
           stxEvents: [],
           ftEvents: [],
-          nftEvents: [],
+          nftEvents: [
+            {
+              event_type: DbEventTypeId.NonFungibleTokenAsset,
+              asset_event_type_id: DbAssetEventTypeId.Mint,
+              value: bnsNameCV('xyz.abc'),
+              asset_identifier: 'SP000000000000000000002Q6VF78.bns::names',
+              recipient: 'ST5RRX0K27GW0SP3GJCEMHD95TQGJMKB7G9Y0X1ZA',
+              tx_id: tx3.tx_id,
+              tx_index: tx3.tx_index,
+              block_height: block2b.block_height,
+              event_index: 0,
+              canonical: true,
+            },
+          ],
           contractLogEvents: [],
           smartContracts: [contract1],
           names: [
             {
-              name: 'xyz',
+              name: 'xyz.abc',
               address: 'ST5RRX0K27GW0SP3GJCEMHD95TQGJMKB7G9Y0X1ZA',
               namespace_id: 'abc',
-              registered_at: 2,
+              registered_at: block2b.block_height,
               expire_block: 14,
               zonefile:
                 '$ORIGIN muneeb.id\n$TTL 3600\n_http._tcp IN URI 10 1 "https://blockstack.s3.amazonaws.com/muneeb.id"\n',
               zonefile_hash: 'b100a68235244b012854a95f9114695679002af9',
               canonical: true,
-              tx_id: tx2.tx_id,
-              tx_index: tx2.tx_index,
+              tx_id: tx3.tx_id,
+              tx_index: tx3.tx_index,
             },
           ],
           namespaces: [
@@ -3936,8 +3966,8 @@ describe('postgres datastore', () => {
               status: 'ready',
               buckets: '1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1',
               canonical: true,
-              tx_id: tx2.tx_id,
-              tx_index: tx2.tx_index,
+              tx_id: tx3.tx_id,
+              tx_index: tx3.tx_index,
             },
           ],
         },
@@ -3985,7 +4015,11 @@ describe('postgres datastore', () => {
     });
     expect(names.results.length).toBe(1);
 
-    name = await db.getName({ name: 'xyz', includeUnanchored: false });
+    name = await db.getName({
+      name: 'xyz.abc',
+      includeUnanchored: false,
+      chainId: ChainID.Mainnet,
+    });
     assert(name.found);
     expect(name.result.canonical).toBe(true);
     expect(name.result.index_block_hash).toBe(block2.index_block_hash);
@@ -4046,7 +4080,11 @@ describe('postgres datastore', () => {
     };
     await db.update({ block: block4b, microblocks: [], minerRewards: [], txs: [] });
 
-    name = await db.getName({ name: 'xyz', includeUnanchored: false });
+    name = await db.getName({
+      name: 'xyz.abc',
+      includeUnanchored: false,
+      chainId: ChainID.Mainnet,
+    });
     assert(name.found);
     expect(name.result.canonical).toBe(true);
     expect(name.result.index_block_hash).toBe(block2b.index_block_hash);
@@ -4821,7 +4859,7 @@ describe('postgres datastore', () => {
       sender_address: 'sender-addr-test',
     };
 
-    const rowCount = await db.updateNFtMetadata(nftMetadata, 0);
+    const rowCount = await db.updateNFtMetadata(nftMetadata, 1);
     expect(rowCount).toBe(1);
 
     const query = await db.getNftMetadata(nftMetadata.contract_id);
@@ -4843,7 +4881,7 @@ describe('postgres datastore', () => {
       sender_address: 'sender-addr-test',
     };
 
-    const rowCount = await db.updateFtMetadata(ftMetadata, 0);
+    const rowCount = await db.updateFtMetadata(ftMetadata, 1);
     expect(rowCount).toBe(1);
 
     const query = await db.getFtMetadata(ftMetadata.contract_id);