+++ src/tests/token-tests.ts
@@ -53,6 +53,7 @@ describe('/extended/v1/tokens tests', () => {
     expect(result1.total).toEqual(1);
     expect(result1.results[0].asset_identifier).toEqual(assetId1);
     expect(result1.results[0].tx_id).toEqual('0x5454');
+    expect(result1.results[0].block_height).toEqual(block1.block.block_height);
 
     // Request: with metadata
     const request2 = await supertest(api.server).get(
@@ -64,6 +65,7 @@ describe('/extended/v1/tokens tests', () => {
     expect(result2.total).toEqual(1);
     expect(result2.results[0].asset_identifier).toEqual(assetId1);
     expect(result2.results[0].tx.tx_id).toEqual('0x5454');
+    expect(result2.results[0].block_height).toEqual(block1.block.block_height);
 
     // Mint another NFT
     const block2 = new TestBlockBuilder({
@@ -91,6 +93,7 @@ describe('/extended/v1/tokens tests', () => {
     expect(result3.total).toEqual(2);
     expect(result3.results[0].asset_identifier).toEqual(assetId2);
     expect(result3.results[0].tx_id).toEqual('0x5464');
+    expect(result3.results[0].block_height).toEqual(block2.block.block_height);
 
     // Request: filtered by asset id
     const request4 = await supertest(api.server).get(
@@ -102,6 +105,7 @@ describe('/extended/v1/tokens tests', () => {
     expect(result4.total).toEqual(1); // 1 result only
     expect(result4.results[0].asset_identifier).toEqual(assetId2);
     expect(result4.results[0].tx_id).toEqual('0x5464');
+    expect(result4.results[0].block_height).toEqual(block2.block.block_height);
 
     // Transfer one NFT from addr1 to addr2
     const block3 = new TestBlockBuilder({
@@ -129,6 +133,7 @@ describe('/extended/v1/tokens tests', () => {
     expect(result5.total).toEqual(1);
     expect(result5.results[0].asset_identifier).toEqual(assetId1);
     expect(result5.results[0].tx_id).toEqual('0x5454');
+    expect(result5.results[0].block_height).toEqual(block1.block.block_height);
 
     // Request: addr2 has the other
     const request6 = await supertest(api.server).get(
@@ -140,6 +145,7 @@ describe('/extended/v1/tokens tests', () => {
     expect(result6.total).toEqual(1);
     expect(result6.results[0].asset_identifier).toEqual(assetId2);
     expect(result6.results[0].tx_id).toEqual('0x5484');
+    expect(result6.results[0].block_height).toEqual(block3.block.block_height);
 
     // Transfer NFT from addr2 to addr3 in microblock
     const microblock1 = new TestMicroblockStreamBuilder()
@@ -177,6 +183,7 @@ describe('/extended/v1/tokens tests', () => {
       block_height: 4,
       index_block_hash: '0x04',
       parent_index_block_hash: '0x03',
+      parent_microblock_hash: '0x11',
     })
       .addTx({ tx_id: '0x5555' })
       .build();
@@ -470,6 +477,7 @@ describe('/extended/v1/tokens tests', () => {
       block_height: 3,
       index_block_hash: '0x03',
       parent_index_block_hash: '0x02',
+      parent_microblock_hash: '0x11',
     })
       .addTx({ tx_id: '0x1004' })
       .build();
@@ -622,6 +630,7 @@ describe('/extended/v1/tokens tests', () => {
       block_height: 7,
       index_block_hash: '0x07',
       parent_index_block_hash: '0x06',
+      parent_microblock_hash: '0x14',
     })
       .addTx({ tx_id: '0x100b' })
       .build();
@@ -795,6 +804,7 @@ describe('/extended/v1/tokens tests', () => {
       block_height: 3,
       index_block_hash: '0x03',
       parent_index_block_hash: '0x02',
+      parent_microblock_hash: '0x11',
     })
       .addTx({ tx_id: '0x1004' })
       .build();
@@ -942,6 +952,7 @@ describe('/extended/v1/tokens tests', () => {
       block_height: 7,
       index_block_hash: '0x07',
       parent_index_block_hash: '0x06',
+      parent_microblock_hash: '0x14',
     })
       .addTx({ tx_id: '0x100b' })
       .build();